# [Enhancement] Improve cache select performance