--- a/be/src/block_cache/kv_cache.h
+++ b/be/src/block_cache/kv_cache.h
@@ -61,6 +61,8 @@ public:
 
     virtual Status read_object(const std::string& key, DataCacheHandle* handle, ReadCacheOptions* options) = 0;
 
+    virtual bool exist(const std::string& key) const = 0;
+
     // Remove data from cache. The offset must be aligned by block size
     virtual Status remove(const std::string& key) = 0;