--- a/be/src/exec/hdfs_scanner_orc.h
+++ b/be/src/exec/hdfs_scanner_orc.h
@@ -17,9 +17,9 @@
 #include <orc/OrcFile.hh>
 
 #include "exec/hdfs_scanner.h"
+#include "formats/disk_range.hpp"
 #include "formats/orc/orc_chunk_reader.h"
 #include "formats/orc/orc_input_stream.h"
-#include "formats/orc/utils.h"
 
 namespace starrocks {