… be/src/exec/CMakeLists.txt
… be/src/exec/hdfs_scanner.h
… be/src/exec/hdfs_scanner_orc.cpp
… be/src/exec/hdfs_scanner_parquet.cpp
… be/src/exec/iceberg/iceberg_delete_builder.cpp
… be/src/exec/iceberg/iceberg_delete_builder.h
… be/src/formats/CMakeLists.txt
… be/src/formats/disk_range.hpp
… be/src/formats/orc/orc_chunk_reader.cpp
… be/src/formats/orc/orc_chunk_reader.h
… be/src/formats/orc/orc_input_stream.h
… be/src/formats/orc/utils.h
… be/src/formats/parquet/file_reader.cpp
… be/src/formats/parquet/file_reader.h
… be/src/io/CMakeLists.txt
… be/src/io/cache_input_stream.h
… be/src/io/cache_select_input_stream.hpp
… be/test/CMakeLists.txt
… be/test/exec/hdfs_scanner_test.cpp
… be/test/exec/iceberg/iceberg_delete_builder_test.cpp
… be/test/formats/disk_range_test.cpp
… be/test/formats/orc/utils_test.cpp
… be/test/formats/parquet/column_converter_test.cpp
… be/test/formats/parquet/file_reader_test.cpp
… be/test/formats/parquet/file_writer_test.cpp
… be/test/formats/parquet/iceberg_schema_evolution_file_reader_test.cpp
… be/test/formats/parquet/page_index_test.cpp
… be/test/formats/parquet/parquet_cli_reader.h
… be/test/formats/parquet/parquet_file_writer_test.cpp
… be/test/formats/parquet/parquet_footer_test.cpp