## Why I'm doing:
Original cache select is reuse query process, it will require high CPU & memory and is also slow. 

## What I'm doing:
For textfile: download it directly, and write it to DataCache
For orc/parquet: Read the footer first, then load the selected column's data into DataCache

Propose `DiskRange` concept to avoid coupling with IORange in SharedBufferInputStream. IORange is an inner class in SharedBufferInputStream, shouldn't use it outside.

## Performance report

First read(Result is not 100% correctly, because of network jitter)

| Table      | Previous Implementataion | New Implemention |
| ----------- | ----------- | ------------ |
| 2.2G TEXT 100Gtpch customer      | 10s       | 8.5s | 
| 25.2G Parquet 100Gtpch lineitem   | 45.87s        | 16.8s | 
| 15.7G ORC 100G tpch lineitem | 20.01s | 18.48s | 
| 5.8MB, 50,000 iceberg delete files | 13.06s | 2.05s |

Assume all data is already cached in DataCache

| Table      | Previous Implementataion | New Implemention |
| ----------- | ----------- | ------------ |
| 2.2G TEXT 100Gtpch customer      | 1.46s       | 0.02s | 
| 25.2G Parquet 100Gtpch lineitem   | 6.52s        | 0.09s | 
| 15.7G ORC 100G tpch lineitem | 7.3s | 0.41s | 
| 5.8MB, 50,000 iceberg delete files | 11.7s | 0.22s |


## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5